@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
-    "start": "ng serve --delete-output-path=false",
+    "start": "ng serve",
     "build": "node build.js",
     "test": "karma start",
     "install-package": "npm run build && npm pack ./dist && npm install ngx-ui-scroll-0.0.5-rc.2.tgz --no-save",
@@ -12,52 +12,52 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.7",
-    "@angular/common": "^5.2.7",
-    "@angular/core": "^5.2.7",
-    "@angular/forms": "^5.2.7",
-    "@angular/http": "^5.2.7",
-    "@angular/platform-browser": "^5.2.7",
-    "@angular/platform-browser-dynamic": "^5.2.7",
-    "@angular/router": "^5.2.7",
-    "core-js": "^2.5.3",
-    "ngx-bootstrap": "^2.0.3",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.20"
+    "@angular/animations": "^6.0.4",
+    "@angular/common": "^6.0.4",
+    "@angular/core": "^6.0.4",
+    "@angular/forms": "^6.0.4",
+    "@angular/http": "^6.0.4",
+    "@angular/platform-browser": "^6.0.4",
+    "@angular/platform-browser-dynamic": "^6.0.4",
+    "@angular/router": "^6.0.4",
+    "core-js": "^2.5.7",
+    "ngx-bootstrap": "^3.0.0",
+    "rxjs": "^6.2.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.3",
-    "@angular/compiler": "^5.2.7",
-    "@angular/compiler-cli": "^5.2.7",
-    "@angular/language-service": "^5.2.7",
-    "@types/jasmine": "^2.5.54",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^6.0.95",
-    "chalk": "^2.3.0",
-    "codelyzer": "^4.3.0",
+    "@angular-devkit/build-angular": "^0.6.8",
+    "@angular/cli": "^6.0.7",
+    "@angular/compiler": "^6.0.4",
+    "@angular/compiler-cli": "^6.0.4",
+    "@angular/language-service": "^6.0.4",
+    "@types/jasmine": "^2.8.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^6.0.112",
+    "chalk": "^2.4.1",
+    "codelyzer": "~4.2.1",
     "cpx": "^1.5.0",
-    "jasmine-core": "~2.6.2",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
+    "jasmine-core": "~2.99.1",
+    "karma": "~1.7.1",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.3.1",
+    "karma-coverage-istanbul-reporter": "^2.0.0",
     "karma-firefox-launcher": "^1.1.0",
-    "karma-ie-launcher": "^1.0.0",
-    "karma-jasmine": "^1.1.1",
+    "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-spec-reporter": "0.0.32",
-    "karma-webpack": "^2.0.9",
+    "karma-webpack": "^4.0.0-beta.0",
     "renamer": "^0.6.1",
     "rollup": "^0.53.4",
     "rollup-plugin-license": "^0.6.0",
     "rollup-plugin-node-resolve": "^3.0.2",
     "rollup-plugin-sourcemaps": "^0.4.2",
     "shelljs": "^0.7.8",
-    "ts-loader": "^3.2.0",
-    "ts-node": "~3.2.0",
-    "tslint": "^5.9.1",
+    "ts-loader": "^4.3.1",
+    "ts-node": "~5.0.1",
+    "tslint": "~5.9.1",
     "typescript": "2.7.2",
-    "uglify-js": "^3.3.7"
+    "uglify-js": "^3.4.0"
   }
 }
