@@ -1,5 +1,4 @@
-import { Observable } from 'rxjs/Observable';
-import { Subscription } from 'rxjs/Subscription';
+import { Observable, Subscription } from 'rxjs';
 
 import { Datasource } from './interfaces/index';
 import { Settings } from './classes/settings';
@@ -19,6 +18,7 @@ export class Scroller {
   readonly callWorkflow: Function;
   private logs: Array<any> = [];
 
+  public version: string;
   public datasource: Datasource;
   public settings: Settings;
   public routines: Routines;
@@ -33,6 +33,7 @@ export class Scroller {
   constructor(context, callWorkflow: Function) {
     // this._bindData = () => context.changeDetector.markForCheck();
     this._bindData = () => context.changeDetector.detectChanges();
+    this.version = context.version;
     this.datasource = checkDatasource(context.datasource);
     this.callWorkflow = callWorkflow;
 
@@ -41,7 +42,7 @@ export class Scroller {
     this.viewport = new Viewport(context.elementRef, this.settings, this.routines);
     this.buffer = new Buffer();
     this.state = new State();
-    this.adapter = new Adapter(this.callWorkflow);
+    this.adapter = new Adapter(this);
 
     this.datasource.adapter = this.adapter;
     this.cycleSubscriptions = [];
